Course completed

+++ backend/src/controllers/notes_controller.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import Note from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const index: RequestHandler = async (req, res, next) => {
+    const authUserId = req.session.userId;
     try {
-        const notes = await Note.find().exec();
+        assertIsDefined(authUserId);
+        const notes = await Note.find({userId: authUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -14,13 +17,18 @@ export const index: RequestHandler = async (req, res, next) => {
 
 export const show: RequestHandler = async (req, res, next) => {
     const id = req.params.id;
+    const authUserId = req.session.userId;
     try {
+        assertIsDefined(authUserId);
         if (!mongoose.isValidObjectId(id)){
             throw createHttpError(400, "Invalid Id")
         }
         const note = await Note.findById(id).exec();
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+        if (!note.userId.equals(authUserId)) {
+            throw createHttpError(401, "User is not authorized to access this note");
         }
         res.status(200).json(note)
     } catch (error) {
@@ -35,13 +43,16 @@ interface CreateBody {
 
 export const create: RequestHandler<unknown, unknown, CreateBody, unknown> = async (req, res, next) => {
     const { title, text } = req.body;
+    const authUserId = req.session.userId;
     try {
+        assertIsDefined(authUserId);
         if (!title) {
-            throw createHttpError(400, "Note must have a title")
+            throw createHttpError(400, "Note must have a title");
         }
         const note = await Note.create({
             title,
-            text
+            text,
+            userId: authUserId
         });
         res.status(201).json(note);
     } catch (error) {
@@ -61,7 +72,9 @@ interface UpdateBody {
 export const update: RequestHandler<UpdateParams, unknown, UpdateBody, unknown> = async(req, res, next) => {
     const { title, text } = req.body;
     const id = req.params.id;
+    const authUserId = req.session.userId;
     try {
+        assertIsDefined(authUserId);
         if (!mongoose.isValidObjectId(id)){
             throw createHttpError(400, "Invalid Id")
         }
@@ -72,9 +85,13 @@ export const update: RequestHandler<UpdateParams, unknown, UpdateBody, unknown> 
         if (!note) {
             throw createHttpError(404, "Note not found");
         }
+        if (!note.userId.equals(authUserId)) {
+            throw createHttpError(401, "User is not authorized to access this note");
+        }
 
         note.title = title;
-        note.text = text
+        note.text = text;
+        note.userId = authUserId;
         
         const newNote = await note.save();
         res.status(200).json(newNote);
@@ -85,13 +102,18 @@ export const update: RequestHandler<UpdateParams, unknown, UpdateBody, unknown> 
 
 export const destroy: RequestHandler = async (req, res, next) => {
     const id = req.params.id;
+    const authUserId = req.session.userId;
     try {
+        assertIsDefined(authUserId);
         if (!mongoose.isValidObjectId(id)){
             throw createHttpError(400, "Invalid Id")
         }
         const note = await Note.findById(id).exec();
         if (!note) {
             throw createHttpError(404, "Note not found");
+        }
+        if (!note.userId.equals(authUserId)) {
+            throw createHttpError(401, "User is not authorized to delete this note");
         }
         await note.deleteOne();
         res.sendStatus(204);
